requirements.txt
fastapi
requests
python-dotenv
openai
uvicorn
